fixed test since trailing / removal. sshcommand option
Previous commit removed forced trailing slash (/) but PREPARE step
still checked existence of slash in repo= value -- and slashless
path made it fail (hardly relevant in real usage, but still broken).
Now just emptiness of repo= value is checked.

Added sshcommand= configuration option -- when given GIT_SSH_COMMAND
is set to the value so that when ssh remote is used it is in effect.