change "relative" repo 'path' to '../path' when no /'s in 'path'
In case of git-aafs it is more convenient that plain slashless
string as repo resolves to url that is in same "directory level"
as the git repository these tools are run on. It seems like
sensible decision to lay out git-aafs repository alongside other
repositories in the same location.

For git-annex-remote-git-aafs.pl this does not matter (as repo path
is given in 'initremote' just once) but is changed for consistency.